+++ src/templates/json/overrides/wilson-ecommerce--wilson.js
@@ -29,7 +29,8 @@ export default class extends JSONTemplate {
    * @param {Variant} [variant]
    */
   constructMPN(variant) {
-    return variant ? variant.sku : this.product.sku;
+    // only include mpn on product-level
+    return variant ? undefined : this.product.sku;
   }
 
   renderOffers() {